requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
filelock==3.17.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
passlib==1.7.4
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.4.0
rsa==4.9.1
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
uvicorn==0.34.0
